+++ src/dashboard/dashboard.styled.ts
@@ -1,8 +1,10 @@
 import styled from 'styled-components/native'
+import { Host } from 'react-native-portalize'
 import MaterialCommunityIcons from 'react-native-vector-icons/MaterialCommunityIcons'
 import { COLOR } from '../core/colors/colors.constants'
 import { TEXT_SIZE } from '../core/text/text.constants'
 import { SafeAreaView } from 'react-native-safe-area-context'
+import { FocusAwareStatusBar } from '../core/focus-aware-status-bar/focus-aware-status-bar'
 
 export const SearchButtonIcon = styled(MaterialCommunityIcons).attrs({
   name: 'book-search',
@@ -31,4 +33,14 @@ export const DashboardButtonsContainer = styled.View`
   flex-direction: row;
   justify-content: space-around;
   padding-bottom: 10px;
+`
+
+export const DashboardStatusBar = styled(FocusAwareStatusBar).attrs({
+  backgroundColor: COLOR.WHITE,
+  animated: true,
+  barStyle: 'dark-content',
+})``
+
+export const DashboardHost = styled(Host)`
+  background-color: ${COLOR.WHITE};
 `