+++ src/dictionary/dictionary.styled.ts
@@ -2,6 +2,7 @@ import styled from 'styled-components/native'
 import { SafeAreaView } from 'react-native-safe-area-context'
 import { COLOR } from '../core/colors/colors.constants'
 import { CustomTextInput } from '../core/custom-text-input/custom-text-input'
+import { FocusAwareStatusBar } from '../core/focus-aware-status-bar/focus-aware-status-bar'
 
 export const DictionarySafeAreaContainer = styled(SafeAreaView)`
   flex: 1;
@@ -17,6 +18,8 @@ export const DictionaryTextInput = styled(CustomTextInput).attrs({
   maxLength: 15,
 })``
 
-export const DictionaryStatusBar = styled.StatusBar.attrs({
+export const DictionaryStatusBar = styled(FocusAwareStatusBar).attrs({
   backgroundColor: COLOR.DODGER_BLUE,
+  animated: true,
+  barStyle: 'light-content',
 })``